🦄 refactor(cli): Renamed command
Renamed "create" command in cli to "add"
