+++ packages/cli/src/commands/add.ts
